Juan (#8)
Juan,
Thank you for sprucing up the code in exploreMap.js based on best practices (commenting and indentation tips, simplifying var declarations by getting rid of empty object or array, extracting functions so functions not coupled with DOM elements or other functions...). I agree about removing the beginnings of the Expert function. I did so when I started my refactor branch, which is specifically for refactoring exploreMap.js. It was good that you forked from master because, as of today (April 23, 2016), the refactor branch only contains the revealing module approach, despite that there's no need to privatize that data; however, if it were the case that it needed to be private and that pattern was practical to use, I did manage to pass the productTypes values outside of the module (i.e., function ParseJsonModule() ) by storing them in var i for each iteration through the contacts array.  I'll likely start a new branch based off your changes. That branch will be for modifying exploreMap.js in order to show just one state expert if the clicked state contains an expert for the productType selected in the dropdown.